requirements.txt
pygame-ce==2.5.2
pytmx==3.32